Update file structure and rename
- Rename project defend_data_capture to match the repo

- Moved apps up a level in the file structure to match typical django apps

- Rename supply_chain_updates to just be supply_chains
