--- api/pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE = api.settings
