+++ defend_data_capture/accounts/api_views.py
@@ -1,7 +1,7 @@
-from api.accounts.models import User
+from accounts.models import User
 from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated
-from api.accounts.serializers import UserSerializer
+from accounts.serializers import UserSerializer
 
 
 class UserViewSet(viewsets.ModelViewSet):
