+++ defend_data_capture/accounts/serializers.py
@@ -1,4 +1,4 @@
-from api.accounts.models import User
+from accounts.models import User
 from rest_framework import serializers
