+++ defend_data_capture/defend_data_capture/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for api project.
+ASGI config for defend_data_capture project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "api.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "defend_data_capture.settings")
 
 application = get_asgi_application()
