+++ defend_data_capture/defend_data_capture/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for api project.
+WSGI config for defend_data_capture project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "api.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "defend_data_capture.settings")
 
 application = get_wsgi_application()
