+++ defend_data_capture/pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = defend_data_capture.settings
