+++ defend_data_capture/supply_chains/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class SupplyChainUpdateConfig(AppConfig):
-    name = "supply_chain_update"
+    name = "supply_chains"
